pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ctxpact"
version = "0.1.0"
description = "Resilient context compaction gateway for local LLM inference"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Manoj", email = "manojbhat.research@gmail.com" }]
keywords = ["llm", "context", "compaction", "proxy", "mlx", "vllm"]

dependencies = [
    "fastapi>=0.115.0,<0.129.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0,<0.28.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "pyyaml>=6.0.0",
    "tiktoken>=0.7.0",
    "aiosqlite>=0.20.0",
    "sse-starlette>=2.1.0",
    "dspy>=3.1.1",
    "chromadb>=1.0.0",
    "sentence-transformers>=3.0.0",
    "llmlingua>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "respx>=0.21.0",
]

[project.scripts]
ctxpact = "ctxpact.server:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
